Add SettingsDialog component and update card themes to include Casino
