+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export { CardComponent } from './CardComponent';
 export { CardStyleCustomizer } from './CardStyleCustomizer';
+export { SettingsDialog } from './SettingsDialog';
